pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "udaplay"
version = "0.1.0"
description = "UdaPlay AI research agent for video game Q&A"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "chromadb>=1.0.4",
  "openai>=1.73.0",
  "pydantic>=2.11.3",
  "python-dotenv>=1.1.0",
  "tavily-python>=0.5.4"
]

[tool.setuptools]
packages = ["udaplay"]

[project.scripts]
udaplay = "udaplay.main:main"
